api/requirements.txt
azure-functions
azure-servicebus
azure-cosmos
azure-identity
